useSafeIntl hook + better code
