+++ hat/assets/js/apps/Iaso/hooks/intl.js
@@ -1,0 +1,4 @@
+import { useIntl } from 'react-intl';
+import { patchIntl } from '../libs/intl/utils';
+
+export const useSafeIntl = () => patchIntl(useIntl());
